Split the histogram class into its own file, greatly robustified the numerics, and made it a lot more flexible and memory-consuming <sigh>. This is to help make progress on the central-limit scheme, where we still have a poor idea of what the zscore distributions look like. The histogram class is flexible enough to give us nice pictures of that now.
Note that new min, max, and median statistics are displayed for all
histograms now (and computing percentile cutoffs would be easy to add).

Note that if you have a histogram object, you can now pass the # of
buckets desired to the display() method (no code exploits that yet,
but it means, e.g., that if you discover you really wanted more buckets,
you potentially don't have to rerun the test).  HistToGNU.py in
particular may be able to exploit that immediately.