The configuration page now tells you where your ini file is.
It also respects the rules in Options.py for finding the ini file.

I've pulled out all the code for running OptionsConfig.py as a
standalone script (this is part of moving it into the spambayes
package from the scripts area, but I've done it in stages to
minimise lossage through abusing CVS).