Add a refresh button to the review messages page in the pop3proxy web interface.
Add a "check again" link in the "there are no more untrained
messages" page in the pop3proxy web interface.

Fix a bug in UserInterface.py that meant that options in sections
with an underscore in the name (like html_ui) could not be set
via the user interface (note that this is really only a patch, as it
will still fail if there is more than one underscore in the section
name, but let's just not do that, ok? <wink>)

Add an option "html_ui":"display_to" that, if True, will display
the To header (along with from, subject and the classification)
in the pop3proxy review messages web interface.  This is an
option since it makes the page wider than would fix on small
resolutions, and is probably of limited use for those with only
one mail account.