I believe this is the cause of quite a few bug reports recently.
Because we now use whichdb to figure out what type of file
the db is, rather than the open_best function, if we were using
windows and python 2.2 we would try and use dbhash instead
of db3hash, which (as I understand it) is a big bad thing.

Fix this.  This fixes the problem that can be reproduced (before this checkin) like this:

start sb_server.py with fresh databases.
stop sb_server.py
try and start sb_server.py.  you get a 'messageinfo db has no attribute db' traceback.

(Bugfix candidate).