New experimental option: x-use_bigrams.
This is the mixed uni/bi-grams scheme outlined by Tim/Graham.  Basically both uni and
bi grams (of *tokens* not words) are generated and the strongest of the unigrams and
bigram is used, rather than all three.

To get more information:
>>> from spambayes.Options import options
>>> print options.display_full("Classifier", "x-use_bigrams")

Also fix a bug in the web interface where the probability would be incorrectly
calculated on 'show clues'.