Default to twenty search results rather than just one.
Made the search form do a GET rather than a POST.  Previously,
going Back to the search results would show an "Expired" page
(at least in IE6).  It now runs the search again, which is almost as
bad but at least you see the page.  We'll need a proper page expiry
mechanism eventually.