** Incompatible change **
Let's call this 1.0a9, because 7+1 is a bad number, and that makes us match the Outlook plug-in.

In imapfilter, hint at a good imaplib debug level, and don't put the todo list in the docstring.

If the allowed values for an option is a tuple of values, then the type is correct if the value is one
of those types.  Also remove duplicate code from OptionsClass.py

Get all the scripts to determine the database name and type via a utility function in storage.py,
so that everything is consistent across scripts (and it removes some code duplication, although the
options list is looped through twice).

The consistent style is to use "-p FILE" for pickle FILE, and "-d FILE" for dbm FILE.  The two
scripts that already used -p now use -P instead (imapfilter, notesfilter).

Dumped the "mysql::"/"pgsql::" format for electing to use mySQL/pgSQL (I doubt anyone is
using this, anyway).  Replaced it with a proper selection, by making the persistent_use_database
option multi-choice, instead of boolean (it's backwards compatible, so that True/"True" and False/"False"
are "dbm" and "pickle", respectively).  This also provides room for a "zodb" option, at some point
<wink>.

I've tested all the scripts except for notesfilter, which doesn't want to run here, and hammie/hammiebulk.  (By tested, I mean that I've run them far enough to see that they're opening the correct db with -p/-d/none and raising an error with both -p and -d).  That said, testing by day-to-day users of the scripts would be great.

Fixes [ 824628 ] sb_dbexpimp.py, sb_filter.py option incompatibility
Closes [857595] Same argument on the command line
Closes [857597] On command line use -D for pickle
Closes [861656] Missing file for patch