If the special folders can't be found, then report this to the user.
The idea here is to tell people what the problem is when they delete their spam/unsure folder,
rather than just leave the mail unfiltered, and suggest how they could fix it.

The change will be ok as long as nothing else triggers the "Error finding the
MAPI folders for a folder switch event" error.

Note that those with the spam/unsure folder *in* the deleted items folder are still on their own.