Fix [ 906581 ] Assertion failed in search subject
Can an email message's items/headers be non-strings?  It appears from
the bug report that this must be the case, although I'm not sure what would
cause that to be so.  The email package's documentation is very difficult (for me)
to make sense of, so I'm not really sure.

In any case, forcing them to strings won't hurt really, except that a search might
be a bit slower (but that's a rare occurence).