Fix [851785] Pop3proxy stores Ham/Spam/Unsure subject line in message cache
If the notate_to or notate_subject options are used, then the text that gets stored
in the sb_server cache includes these modifications.  Later, when these messages
are used for training, this information is also used (so there is a subject:spam clue
that is introduced by spambayes, for example.  This isn't really ideal (although nor
is modifying these headers in the first place).

This changes the review page to ask the corpus to strip out the classifications from
the to and subject lines if (and only if) the appropriate options are set.

The only flaw I can see with this is that if there are cached messages that were
added *before* the option was set, and so don't have the classification, but do have
an identical one from some other source (an upstream filter, for example), this
information will be stripped.  It doesn't seem likely that this would be common.