Fix the potential open relay problem with smtpproxy.
Fix 1:  The listening_ports option for smtproxy is now of type SERVER rather
than type PORT, as the pop3proxy one already was.  This means that you
can set the listening port to be (for example) localhost:25, and only localhost
connections will be accepted.

Fix 2:

There are two new options, one in each of the pop3proxy and smtpproxy
sections, both called allow_remote_connections.  These do the same thing
as the option of the same name in the html_ui section (it's basically the same
code!).

By default, connections from outside localhost will return a POP3/SMTP
error message and close.  You can explicitly open this up to certain IPs or
to anyone, if you really want to.

Fix 3:

I've added to the smtpproxy option documentation to point out that entering
in your smtp server details isn't necessary if you're not going to use it to train
(although it does also allow the bug report to be sent...)

---

I've tested this as much as I can, using my machine and another machine on
my network.  I'm behind a firewall I don't control, so can't test anything more
remote than that, or use one of the available testing websites to check if this
works properly.  Given that this is a major concern, it would be great if someone
else was able to test this.

This also closes [ 797579 ] Disable connections to POP3 and SMTP from remote hosts