Temporary fix for [ 941596 ] sb_imapfilter.py not adding headers / moving messages
The problem is that not all IMAP servers recognise that the new message is there
immediately after an APPEND, but we need to find it.  A no-op generally gives enough
time for the server to do its thing, and won't hit anyone else significantly.

The proper fix is to put in a loop waiting for the server to dish out the new "EXISTS"
data, but I'll leave that for post 1.0.

Also change the exception catching to catch any exception.  A new message parsing
problem raises a TypeError not an email.Error, so was causing imapfilter to crash.  We
really want to keep going no matter what, otherwise we leave the user's mailbox in an
ugly state.