Psts and the main Exchange store have top level folders with parents (with empty display names), and no grandparents.  Anything below the top level *does* have a grandparent.  This means our test for "top level folder" can be: does it have a parent *and* grandparent.
However, a secondary Exchange account doesn't have the
empty-display-name parent, so the top-level doesn't have a parent, and
the top selectable folder doesn't have a grandparent, and our test
fails.

Allow for this by checking for the "Access denied" exception when getting the
grandparent, and assuming that this means that this is what is happening.
This will only fail if we get an 'access denied' error for the empty-display-name
parent, which should not be the case.

This fixes the "Please select child folder" problem reported by various people
on the list.  I think it closes a SF bug, but the SF site is down, so I can't check.

This would be a good bugfix candidate for 1.0 (or some similar fix, if Mark can
think of a better solution in this case).