Drat.  Using -u on the command line with sb_server hasn't worked for a while (Hmm...I guess the tests should be *run* occasionally, too!).  We need to set these options in the constructor, not in init(), because init gets called after we load in the command line args.
This does mean that any changes to the option values need to be also updated
in state, or a new state has to be made, rather than calling init(), but I believe
we always do this anyway.  In addition, most of these are values that only really
apply to starting up from the command line.

Fixes part of [ 981970 ] tests failing