I emailed spambayes-dev about this on Oct 22, but then clean forgot about checking  the change in!  Thankfully a spambayes@python.org message reminded me, or this bug  would have made it through to 1.0.1...
It appears that our installers aren't offering to create a startup icon for
sb_server users as it should be.

The problem is the "Check: InstallingProxy" line in the Inno script.
Although the selection code has been run by the time the tasks are offered,
it still has the default (False) value.  I've played around with the code,
but can't figure a way around this (although my Pascal is extremely rusty).  Maybe
it's an Inno bug or something.

We can fix it by removing that Check.  Outlook users don't get that page
anyway, so they won't see the option (this is what happens with the desktop
icon).  However, since we want it checked by default, it will appear in the
text box of additional tasks, even though it doesn't happen.  I say this doesn't matter,
and the vast quiet on spambayes-dev tells me people either agree or don't care <wink>.