Mine NNTP-Posting-Host headers.  This is part of an effort to put some SpamBayes smarts into the Mailman gate_news program.
Anecdotal evidence on comp.lang.python suggests that certain posting hosts
(I won't name any names, but the one mentioned heavily starts with a 'g'and
has two 'o's in the middle) are more prone to let spam leak into the system.

More testing is yet to be done, but this seems to be promising.